[ 박수로 ] 리팩토링 : post,calendar, comment, user 관련 redux 모두 제거 후 useSWR 대체
