+++ src/components/post/ReportModal.js
@@ -3,6 +3,7 @@ import { Input, Select, Modal, Button, message } from 'antd';
 import { withRouter } from 'react-router';
 import { PUBLIC_IP } from '../../config';
 import axios from 'axios';
+import { errorMessage } from '../../functions/errorHandling';
 function ReportModal({ type, id, history }) {
   const { Option } = Select;
   const { TextArea } = Input;
@@ -22,42 +23,16 @@ function ReportModal({ type, id, history }) {
           message.success('신고가 완료되었습니다. 감사합니다.');
         })
         .catch((error) => {
-          switch (error.response?.status) {
-            case 401:
-              message.error('로그인이 필요합니다.');
-              history.push('/');
-              break;
-            case 403:
-              message.error('접근 권한이 없습니다');
-              break;
-            case 409:
-              message.error('이미 신고한 게시글입니다.');
-              break;
-            default:
-              break;
-          }
+          errorMessage(error.response?.data.message);
         });
     } else if (type === 'comment') {
       await axios
         .post(`${PUBLIC_IP}/reply/${id}/report`, body)
-        .then((response) => {
+        .then(() => {
           message.success('신고가 완료되었습니다. 감사합니다.');
         })
         .catch((error) => {
-          switch (error.response?.status) {
-            case 401:
-              message.error('로그인이 필요합니다.');
-              history.push('/');
-              break;
-            case 403:
-              message.error('접근 권한이 없습니다');
-              break;
-            case 409:
-              message.error('이미 신고한 댓글입니다.');
-              break;
-            default:
-              break;
-          }
+          errorMessage(error.response?.data.message);
         });
     }
     setBody({ content: 1, detail: '' });
